feat: add webworker for test execution
