+++ apps/@sparrow-web/src/pages/workspace-page/sub-pages/RestExplorer/RestExplorerPage.ViewModel.ts
@@ -2026,126 +2026,10 @@ class RestExplorerViewModel {
   };
 
   private async executeScriptTestcases() {
-    const tests: { name: string; passed: boolean; error?: string }[] = [];
-
-    // minimal chai-like expect (you can replace with a real lib like chai)
-    const expect = (actual: any) => ({
-      to: {
-        equal: (expected: any) => {
-          if (actual !== expected)
-            throw new Error(`Expected ${actual} to equal ${expected}`);
-        },
-        notEqual: (expected: any) => {
-          if (actual === expected)
-            throw new Error(`Expected ${actual} to not equal ${expected}`);
-        },
-        exist: () => {
-          if (actual === undefined || actual === null)
-            throw new Error(`Expected value to exist but got ${actual}`);
-        },
-        notExist: () => {
-          if (actual !== undefined && actual !== null)
-            throw new Error(`Expected value to not exist but got ${actual}`);
-        },
-        be: {
-          a: (type: string) => {
-            if (typeof actual !== type)
-              throw new Error(`Expected type ${type} but got ${typeof actual}`);
-          },
-          true: () => {
-            if (actual !== true)
-              throw new Error(`Expected value to be true but got ${actual}`);
-          },
-          false: () => {
-            if (actual !== false)
-              throw new Error(`Expected value to be false but got ${actual}`);
-          },
-          within: (min: number, max: number) => {
-            if (typeof actual !== "number")
-              throw new Error(`Expected a number but got ${typeof actual}`);
-            if (actual < min || actual > max)
-              throw new Error(
-                `Expected ${actual} to be within ${min} and ${max}`,
-              );
-          },
-          lessThan: (expected: number) => {
-            if (!(typeof actual === "number" && typeof expected === "number"))
-              throw new Error(`Expected numbers for comparison`);
-            if (!(actual < expected))
-              throw new Error(`Expected ${actual} to be less than ${expected}`);
-          },
-          greaterThan: (expected: number) => {
-            if (!(typeof actual === "number" && typeof expected === "number"))
-              throw new Error(`Expected numbers for comparison`);
-            if (!(actual > expected))
-              throw new Error(
-                `Expected ${actual} to be greater than ${expected}`,
-              );
-          },
-          empty: () => {
-            if (
-              (Array.isArray(actual) && actual.length > 0) ||
-              (typeof actual === "string" && actual.trim().length > 0) ||
-              (actual &&
-                typeof actual === "object" &&
-                Object.keys(actual).length > 0)
-            ) {
-              throw new Error(
-                `Expected value to be empty but got ${JSON.stringify(actual)}`,
-              );
-            }
-          },
-          notEmpty: () => {
-            if (
-              (Array.isArray(actual) && actual.length === 0) ||
-              (typeof actual === "string" && actual.trim().length === 0) ||
-              (actual &&
-                typeof actual === "object" &&
-                Object.keys(actual).length === 0)
-            ) {
-              throw new Error(`Expected value to not be empty`);
-            }
-          },
-        },
-        contain: (expected: any) => {
-          if (
-            (typeof actual === "string" || Array.isArray(actual)) &&
-            !actual.includes(expected)
-          ) {
-            throw new Error(`Expected ${actual} to contain ${expected}`);
-          }
-        },
-        notContain: (expected: any) => {
-          if (
-            (typeof actual === "string" || Array.isArray(actual)) &&
-            actual.includes(expected)
-          ) {
-            throw new Error(`Expected ${actual} to not contain ${expected}`);
-          }
-        },
-        beInList: (list: any[]) => {
-          if (!Array.isArray(list))
-            throw new Error(`Expected a list but got ${typeof list}`);
-          if (!list.includes(actual))
-            throw new Error(`Expected ${actual} to be in list ${list}`);
-        },
-        notBeInList: (list: any[]) => {
-          if (!Array.isArray(list))
-            throw new Error(`Expected a list but got ${typeof list}`);
-          if (list.includes(actual))
-            throw new Error(`Expected ${actual} to not be in list ${list}`);
-        },
-        have: {
-          all: {
-            keys: (...keys: string[]) => {
-              const missing = keys.filter((k) => !(k in actual));
-              if (missing.length > 0)
-                throw new Error(`Missing keys: ${missing.join(", ")}`);
-            },
-          },
-        },
-      },
+    const worker = new Worker(new URL("../../../../workers/test-script-worker.ts", import.meta.url), {
+      type: "module",
     });
+    // minimal chai-like expect (you can replace with a real lib like chai)
 
     const progressiveTab = createDeepCopy(this._tab.getValue());
     const javaScriptTestCases =
@@ -2156,66 +2040,38 @@ class RestExplorerViewModel {
       if (r) {
         r.response.testResults = [];
         r.response.testMessage = "";
-        // sp object similar to pm
-        const sp = {
-          response: {
-            statusCode: Number(r?.response?.status.split(" ")[0]),
-            body: {
-              text: () => {
-                try {
-                  return r?.response?.body;
-                } catch {
-                  return {};
-                }
-              },
-              json: () => {
-                try {
-                  return JSON.parse(r?.response?.body);
-                } catch {
-                  return {};
-                }
-              },
-            },
-            headers: r?.response?.headers.reduce((acc, h) => {
-              acc[h.key] = h.value;
-              return acc;
-            }, {}),
-            size: r?.response?.size,
-            time: r?.response?.time,
-          },
-          test: (name: string, fn: Function) => {
-            try {
-              fn();
-              tests.push({ name, passed: true });
-            } catch (err: any) {
-              tests.push({ name, passed: false, error: err.message });
-            }
-          },
-          xmlToJSON: (xml: string) => {
-            const parser = new XMLParser();
-            const json = parser.parse(xml);
-            return json;
-          },
-          expect,
-        };
-
-        try {
-          const fn = new Function("sp", javaScriptTestCases);
-          fn(sp); // execute user script with "sp"
-          r.response.testResults = tests.map((t) => ({
-            testId: "", // No ID in script mode
-            testName: t.name,
-            testStatus: t.passed,
-            testMessage: t.error || "",
-          }));
-        } catch (err: any) {
-          console.log(err);
-          r.response.testMessage = `${err.name} - ${err.message}`;
-        }
         restApiDataMap.set(progressiveTab.tabId, r);
+        worker.postMessage({
+          javaScriptTestCases,
+          response: r.response,
+        });
       }
       return restApiDataMap;
     });
+
+    worker.onmessage = (e) => {
+      const { success, tests, error } = e.data;
+      restExplorerDataStore.update((restApiDataMap) => {
+        const r = restApiDataMap.get(progressiveTab?.tabId);
+        if(r){
+          if (success) {      
+            r.response.testResults = tests.map((t) => ({
+              testId: "",
+              testName: t.name,
+              testStatus: t.passed,
+              testMessage: t.error || "",
+            }));
+          } else {
+            r.response.testMessage = error;
+          }
+  
+          restApiDataMap.set(progressiveTab.tabId, r);
+
+        }  
+        return restApiDataMap;
+      });
+      worker.terminate(); // cleanup
+    };
   }
 
   /**
